+++ plugin/cmd/cmdAdd.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"fmt"
 	"log/slog"
+	"net"
 
 	"github.com/bartlettc22/gopher-cni/pkg/cni"
 	"github.com/bartlettc22/gopher-cni/pkg/kubernetes"
@@ -16,8 +17,8 @@ import (
 
 // Overrideable functions for unit tests
 var newKubeClient = kubernetes.NewClientFromConfigFile
-var setupViaHost func(string, string, *wgtypes.Config, *netlink.Addr, *slog.Logger) error = setupWGViaHost
-var setupViaContainer func(string, string, *wgtypes.Config, *netlink.Addr, *slog.Logger) error = setupWGViaContainer
+var setupViaHost func(string, string, *wgtypes.Config, *netlink.Addr, []*net.IPNet, *slog.Logger) error = setupWGViaHost
+var setupViaContainer func(string, string, *wgtypes.Config, *netlink.Addr, []*net.IPNet, *slog.Logger) error = setupWGViaContainer
 
 func Add(args *skel.CmdArgs) (err error) {
 
@@ -72,13 +73,18 @@ func Add(args *skel.CmdArgs) (err error) {
 			return e(log, "failed to parse wireguard address", err)
 		}
 
+		splitTunnelCIDRs, err := k8s.SplitTunnelCIDRs()
+		if err != nil {
+			return e(log, "failed to parse split-tunnel CIDRs", err)
+		}
+
 		switch k8s.CNIMode() {
 		case cni.CNIModeHostOrigin:
-			if err := setupViaHost(args.Netns, cni.InterfaceName, wgConfig.WGConfig, wgAddr, log); err != nil {
+			if err := setupViaHost(args.Netns, cni.InterfaceName, wgConfig.WGConfig, wgAddr, splitTunnelCIDRs, log); err != nil {
 				return e(log, "failed to setup network (host-origin)", err)
 			}
 		case cni.CNIModePodOrigin:
-			if err := setupViaContainer(args.Netns, cni.InterfaceName, wgConfig.WGConfig, wgAddr, log); err != nil {
+			if err := setupViaContainer(args.Netns, cni.InterfaceName, wgConfig.WGConfig, wgAddr, splitTunnelCIDRs, log); err != nil {
 				return e(log, "failed to setup network (pod-origin)", err)
 			}
 		default: